ark_learning_agent/requirements.txt
google-adk
python-dotenv 
mcp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pygments
firebase-admin
requests